Refactored more !
